Better Interface/requirements.txt
setuptools
pyusb
json
pprint
time
npyscreen
